pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Agent OS scaffold for worker flows and media automation"
authors = [{ name = "alex" }]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "google-generativeai>=0.8.3",
    "httpx>=0.27.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.3.0",
    "uvicorn>=0.30.0",
    "typer>=0.12.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
